added pagination and correted fetch order by user id bug

+++ src/main/java/com/grocex_api/ordersService/models/Order.java
@@ -14,6 +14,6 @@ public class Order {
     @GeneratedValue(strategy = GenerationType.UUID)
     private UUID id;
     private UUID customerId;
-    private int totalPrice;
+    private float totalPrice;
     private String status;
 }
